RDP: Advertise TCP-ACTIVE server reflexive candidates
Use UDP STUN requests to discover the external NAT address of the MCU and construct a TCP-ACTIVE server reflexive candidate from it. This will cause the Lync client to add permissions to it's relay candidates that allow the MCU to establish a TCP connection to the edge server.

This should allow RDP to traverse the firewall as long as the following conditions are met

1. The NAT through which the MCU connects assigns the same IP address to UDP and TCP traffic.
2. Any firewalls are configured such that the MCU can establish an outbound TCP connection from an ephemeral port to the external address of the edge server that the Lync client is using.

Note: for RDP calls to traverse the firewall the MCU needs to be configured with a TURN server. However it will not allocate any relays for an RDP call, it simply uses the TURN server as a STUN server.