libnice: Only skip connectivity checks for TCP-PASSIVE candidates when talking to Lync
When Lync is controller it stops responding to connectivity checks as soon as it gets a
response for TCP-ACTIVE candidates. Since it won't respond to our subsequence connectivity
checks we go straight to the discovered state which triggers the component to go ready.

We don't want this behaviour when we are the ICE controller as it will cause the component
state READY event to be signalled before the new-selected-pair event which will mess up the
SIP re-INVITE on ICE complete.