libnice: Fix handling of stored connectivity checks for TCP.
Previously the handling of stored checks was keyed from the current list of
candidate pairs for a given stream/component. Since local TCP passive candidates
are not allowed to form part of connectivity check pairs it is possible for an agent
that only has TCP-passive candidates to have an empty connectivity check list even when
all remote candidates have been set. Because of this the stored check list may never be
processed and the component state never set to ready when the controlling agent has only
TCP passive candidates.