libnice: Restart the connectivity check timer when adding connectivity check pairs.
For an agent with only TCP passive candidates there are initially no check pairs created which
was causing the timer to be stopped. When peer-reflexive candidates were later added they never
became active because the timer routing never ran.

This behaviour was fine when in controlled mode (as is always the case for RDP calls) but would
fail when libnice was the controller.