Skype: Fix media for audio calls.
Fix libnice bug when pruning streams. When we place an outbound call to Skype it always port rejects the video stream causing the MCU to prune the stream within libnice. Unfortunately libnice would then proceed to delete the checklists for all streams, causing ICE to fail for the audio line as well.