libnice: Limit maximum size of TCP transmit queue.
Add a property to the nice agent to control the maximum size of the TCP transmit queue.
Limit the amount of tx buffering in the kernel.

libnice: Reduce default tcp queue size to 100 packets

Enough to be safe for RDP negotiation but stops us queueing too much for real RTP streams.