libnice: Fix Skype interop
To work with latest Skype client we need to

a) Send an MS-IMPLEMENTATION-VERSION stun attribute with a version of 2 (as sent by Lync 2010/2013 clients)
b) Always send binding requests using the standard CRC32 table, not the stupid broken Microsoft CRC32 table. This will probably break interop with OCS 2007 R2 (which we don't support anyway).

Fixes #4428