Apply diff of upstream and imported code
libnice was unfortunately not cleanly imported into main repository
back in the days. It contained auto generated code and was missing
some files.

This patch applies the diff between the upstream version 0.1.3 and the
imported code so that following patches applies cleanly and automatically
when libnice is extracted into separate repository.