gstnicesrc: Always query downstream on negotiate
Default negotiation skips querying caps downstream if the template caps
is any, in which case a caps event will not be sent. A caps event is
required for some elements that does internal configuration based on the
caps. An alternative approach was to implement a caps property, but it
seems preferable to have the elements negotiate the caps instead of
requiring it to be set from the application.