libnice: Protect callback data on TCP sockets
Protect against the socket being destroyed by another thread whilst the mainloop is still running.
If g_source_destroy is called from a thread outside the main loop then there is no guarantee that the
callback will not be invoked again. With the new per agent locking in libnice this meant that we could
end up using an agent pointer that had already been freed when the socket was closed.

Add a separate struct attached to the callback itself and release it in the callbacks GDestroyNotify
func.