gstnicesrc: fix deadlock
It would go something like this:

The element goes to READY, and the srcpad is deactivated, waiting to
join the task from basesrc.

The code has done a unlock, nicely adding a idlesrc on the mainloop
to close that down, but after unlock happens lock, which again would
allow the mainloop to start again, with the create() call inside nicesrc
calling main_loop_run(). This would then go on to block the join
of that task forever.

The solution is to add another variabel, "running", that will be FALSE when
the element is on its way down, so that when the main-loop quits, it will
not be allowed to start up again, since there is now a check for
"running", not just for "unlocked".